Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 #include<math.h>
 int main(){
-    int a,b,c;
+    double a,b,c;
     scanf("%d %d",&a,&b);
     c=pow(b,2);
     if(c==a){
